# Optimise ansi-diff more