Use lower-bound approximation of inter-line edit-distance, making diff-chunk coloring even faster.

The benchmark results (the benchmarks are just two examples, but hopefully they show a trend):
- `th` with has multiple lines, the `ansi-diff` doesn't add much on top of just `Diff`, and that's all is magnitude faster then (character based) levenshtein distance.
- `hs` is a single line, so we end up doing `edit-distance`, but in this particular example `edit-distance` seems to be faster than `text-metrics` (bitfiddling-Myers vs. Wagner-Fisher)

Having `ansi-diff` produce outputs in milliseconds range is great.

```
All
  th
    edit-distance: OK
      170  ms ± 6.6 ms
    text-metrics:  OK
      61.3 ms ± 3.1 ms
    Diff:          OK
      2.18 ms ± 118 μs
    ansidiff:      OK
      2.89 ms ± 231 μs
  hs
    edit-distance: OK
      4.30 ms ± 182 μs
    text-metrics:  OK
      42.1 ms ± 3.1 ms
    Diff:          OK
      899  μs ±  43 μs
    ansidiff:      OK
      5.49 ms ± 375 μs
```